% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.vfeatures.R
\name{assign.vfeatures}
\alias{assign.vfeatures}
\title{assign font, font color, background color to the label of a node}
\usage{
assign.vfeatures(line, DF)
}
\arguments{
\item{dataframe}{with the features to add}
}
\value{
a new line edited according to features
}
\description{
There are two cases: A and B. Example of a line is "77225 'p_524, p_650' x=-87 y=377 lc=0 255 0,"
(A) when there is only one label associate to a vertex_id the cases are only 2:
    (1) the vlab associated in the map.df is not an empty character
    (2) the vlab associated in the map.df is an empty character
for:
    (1) the syntax is VERTEXID VLAB X Y F LC LK
    (2) the syntax is VERTEXID VLAB X Y F LC
(B) when there are more than one label associate to a vertex_id the cases are:
   (1) vlabs are all empty
   (2) only one is not empty
   (3) more than one are not empty
       (3.1) the f, LC and LK are all equal
       (3.2) the f, LC and LK are not-all equal
}
