% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_nexus.R
\name{custom_nexus}
\alias{custom_nexus}
\title{Color SplitsTree dendrograms}
\usage{
custom_nexus(nexus.file, tips, colors, outfile, fromBIGSdb = FALSE, fg,
  vlabels, w = 20, h = 20, s = "r", f, lc, lk, edgecolor = "gray", plot,
  SplitsTree.exe = "/Applications/SplitsTree/SplitsTree")
}
\arguments{
\item{tips}{(leaves of the dendrogram)}

\item{colors}{vector of the same length of the tips indicating the color of each tip}

\item{vlabels}{vector of labels to be plot; leave empty character "" to hide the label.}

\item{w}{width of the tip. If you indicate a single number it will be recycled. Either provide a vector of widths}

\item{h}{height of the tip. Same for w.}

\item{s}{shape of the tip "o" is oval. The other option is "r" rectangle.}

\item{f}{fonts of the labels. Like "Dialog-BOLD-16" or 'Courier-PLAIN-24'. Other option are: Calibri, Arial, Times. For bold -BOLD and not -PLAIN}

\item{lc}{vector of label colors}

\item{lk}{vector of background of the label}

\item{plot}{if TRUE SplitsTree is launched and the nexus displayed}

\item{Splitstree.exe}{path to the executable of SplitsTree}
}
\value{
a new nexus file with colors
}
\description{
Reads a nexus file and adds colors to the leaves
}
\examples{
inputfile <- file.path(path.package('microbio'),'input.nexus')
outfile <- file.path(path.package('microbio'),'output.nexus')
tips <- c("A","R","G","B")
col <- c("red","blue","green","yellow")
s <- c("r","o","r","o")
ws <- c(10,20,30,40)
hs <- c(10,20,30,40)
fg <- c("blue","orange", "violet", "gray8")
lc <- c("gray", "cyan", "pink", "blue")
lk <- c("black", "blue", "black", "yellow")
edgecolor <- "blue"
f <- c("Arial-BOLD-20", "Calibri-BOLD-16", "Times-20", "Dialog-BOLDITALIC-25")
custom_nexus(nexus.file=inputfile, tips=tips, colors=col, vlabels=tips, w=ws, f=f, h=hs, lc=lc, lk=lk, fg= fg, s=s, edgecolor=edgecolor, outfile=outfile, plot=TRUE, SplitsTree.exe = "/Applications/SplitsTree/SplitsTree")
}
