% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.nexus.tips.R
\name{get.nexus.tips}
\alias{get.nexus.tips}
\title{Get tips from dendrogram}
\usage{
get.nexus.tips(nexus.file, fromBIGSdb = FALSE)
}
\value{
a dataframe with tips and verteces index
}
\description{
Reads a nexus file and returns the tips
}
\examples{
inputfile <- file.path(path.package('microbio'),'input.nexus')
get.nexus.tips(nexus.file=inputfile)
}
