% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.tip.colors.R
\name{get.tip.colors}
\alias{get.tip.colors}
\title{Assign colors to tips}
\usage{
get.tip.colors(tree, tip2category, na.col = "black",
  unassigned.col = "gray")
}
\arguments{
\item{tree}{a tree object of class "phylo"}

\item{tip2category}{a named vector. Each entry of the vector is a tip and each name is a level of the catrogory}
}
\value{
a list of two entries: colors with the vector of colors and legend with a vector associating color to levels. Legend is useful to add the legned to the plot.
}
\description{
Given a tree of class "phylo" and a named vector with a category, returns a vector of colors to color the tips on the plots
}
\examples{
require(ape)
x <- c(7, 8, 11, 13, 16, 13, 17, 5, 8, 10, 13,
10, 14, 5, 7, 10, 7, 11, 8, 11, 8, 12,
5, 6, 10, 9, 13, 8)
M <- matrix(0, 8, 8)
M[lower.tri(M)] <- x
M <- t(M)
M [lower.tri(M)] <- x
dimnames(M) <- list(1:8, 1:8)
tr <- nj(M)
tip2category = c(rep(c("hs","mm"),c(4,2)), NA)
names(tip2category) = 1:7
colors = get.tip.colors(tr, tip2category, na.col="black", unassigned.col="gray")[["colors"]]
legenda = get.tip.colors(tr, tip2category, na.col="black", unassigned.col="gray")[["legend"]]
plot(tr, "u", tip.color=colors)
legend("bottomleft", legenda, pch=20, col=names(legenda))
}

