% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_nexus.R
\name{subset_nexus}
\alias{subset_nexus}
\title{Subset a nexus file}
\usage{
subset_nexus(nexus.file, tips, outfile, fromBIGSdb = FALSE)
}
\value{
a nexus file with subset distant matrix
}
\description{
Reads a nexus file and the tips one wants to keep. It create a nexus file with the selected tips
}
