% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.nex.features.R
\name{add.nex.features}
\alias{add.nex.features}
\title{Color SplitsTree dendrograms}
\usage{
add.nex.features(nexus.file, tips, colors, outfile, fromBIGSdb = FALSE,
  fg = "1 1 1", vlabels = "default", w = 20, h = 20, s = "o")
}
\arguments{
\item{tips}{(leaves of the dendrogram)}

\item{w}{width of the tip}

\item{h}{height of the tip}

\item{s}{shape of the tip "o" is oval. The other oprion is "r" rectangle.}

\item{vlabels.}{f='Times New Roman-BOLD-10' lc=204 0 0 lk=0 255 255}

\item{fonts.}{'taxon_X' x=9 y=41 f='Courier-PLAIN-24'. Other option are: Calibri, Arial, Times. For bold -BOLD and not -PLAIN}
}
\value{
a new nexus file with colors
}
\description{
Reads a nexus file and adds colors to the leaves
}
\examples{
inputfile <- file.path(path.package('microbio'),'input.nexus')
outfile <- file.path(path.package('microbio'),'output.nexus')
tips <- c("A","R","G","B")
col <- c("red","blue","green","yellow")
ws <- c(10,20,30,40)
hs <- c(10,20,30,40)
add.nex.features(nexus.file=inputfile, tips=tips, colors=col, w=ws, h=hs, outfile=outfile)
}
