% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.nex.features.R
\name{add.nex.features}
\alias{add.nex.features}
\title{Color SplitsTree dendrograms}
\usage{
add.nex.features(nexus.file, tips, colors, outfile, fromBIGSdb = FALSE,
  fg = "1 1 1", vlabels, w = 20, h = 20, s = "o", f, lc, lk)
}
\arguments{
\item{tips}{(leaves of the dendrogram)}

\item{colors}{vector of the same length of the tips indicating the color of each tip}

\item{vlabels}{vector of labels to be plot; leave empty character "" to hide the label.}

\item{w}{width of the tip. If you indicate a single number it will be recycled. Either provide a vector of widths}

\item{h}{height of the tip. Same for w.}

\item{s}{shape of the tip "o" is oval. The other option is "r" rectangle.}

\item{f}{fonts of the labels. Like "Dialog-BOLD-16" or 'Courier-PLAIN-24'. Other option are: Calibri, Arial, Times. For bold -BOLD and not -PLAIN}

\item{lc}{vector of label colors}

\item{lk}{vector of background of the label}
}
\value{
a new nexus file with colors
}
\description{
Reads a nexus file and adds colors to the leaves
}
\examples{
inputfile <- file.path(path.package('microbio'),'input.nexus')
outfile <- file.path(path.package('microbio'),'output.nexus')
tips <- c("A","R","G","B")
col <- c("red","blue","green","yellow")
ws <- c(10,20,30,40)
hs <- c(10,20,30,40)
add.nex.features(nexus.file=inputfile, tips=tips, colors=col, w=ws, h=hs, outfile=outfile)
}
