% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.nex.features.R
\name{add.nex.features}
\alias{add.nex.features}
\title{Color SplitsTree dendrograms}
\usage{
add.nex.features(nexus.file, tips, colors, outfile, fromBIGSdb = FALSE)
}
\arguments{
\item{tips}{(leaves of the dendrogram)}
}
\value{
a new nexus file with color
}
\description{
Reads a nexus file and adds colors to the leaves
}
\examples{
inputfile <- file.path(path.package('microbio'),'input.nexus')
outfile <- file.path(path.package('microbio'),'output.nexus')
tips <- c("A","R","G","B")
col <- c("red","blue","green","yellow")
add.nex.features(nexus.file=inputfile, tips=tips, colors=col, outfile=outfile)
}

